refactor(ff-stream): extract MuxerCore to eliminate inner-type duplication
Introduce `MuxerCore` in `muxer_core.rs` to own all shared FFmpeg pointer
fields (encoder contexts, scaler, resampler, encoder frames, tracking state)
that were previously duplicated across `RtmpInner`, `SrtInner`,
`LiveHlsInner`, and `LiveDashInner`. Each inner type is now reduced to its
`open_unsafe` (protocol-specific setup) plus delegation to `MuxerCore`.

- Move `pixel_format_to_av` / `sample_format_to_av` into `codec_utils.rs`,
  removing four identical copies.
- Add `close_pb_after_trailer: bool` to `MuxerCore` so that
  `flush_and_close_unsafe` and `free_all` can close `out_ctx.pb` for
  RTMP/SRT (persistent connection) without affecting HLS/DASH (pb already
  closed after header write on Windows).
- Encoder frame allocation is centralised in `MuxerCore::new`; each inner
  `open_unsafe` calls it at the end with its protocol-specific parameters.

Net change: -2071 lines / +164 lines across six files.

Closes #702
